Refactor hybler.js
